Move server code to ai-playground
